feat: makes SocialLinksSidebar sticky within main to avoid footer links duplication
