+++ src/app/globals.css
@@ -47,6 +47,7 @@ body {
 
 main {
   flex-grow: 1;
+  position: relative;
 }
 
 a {
