+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import './globals.css'
 import { Header } from '@/components/Header/Header'
 import { Footer } from '@/components/Footer/Footer'
 import { OverlayMenu } from '@/components/OverlayMenu/OverlayMenu'
-import { SocialLinksSidebar } from '@/app/components/SocialLinksSidebar/SocialLinksSidebar'
+import { SocialLinksSidebar } from '@/components/SocialLinksSidebar/SocialLinksSidebar'
 
 const anton = Anton({
   subsets: ['latin'],
@@ -26,10 +26,12 @@ export default function RootLayout({
     <html lang='en'>
       <body className={`${montserrat.variable} ${anton.variable}`}>
         <Header />
-        {children}
+        <main>
+          {children}
+          <SocialLinksSidebar />
+        </main>
         <Footer />
         <OverlayMenu />
-        <SocialLinksSidebar />
       </body>
     </html>
   )
