+++ src/components/SocialLinksSidebar/SocialLinksSidebar.module.css
@@ -1,0 +1,6 @@
+.root {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+}
